+++ accounts/urls.py
@@ -6,6 +6,7 @@ app_name = "accounts"
 
 urlpatterns = [
     path("login/", views.LoginView.as_view(), name="login"),
+    path("clipper/", views.ClipperLoginView.as_view(), name="clipper"),
     path("logout/", views.LogoutView.as_view(), name="logout"),
     path("create/", views.CreateAccountView.as_view(), name="create"),
     path("update/", views.UpdateAccountView.as_view(), name="update"),