+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Send only one email par orga
 - Send email communicating time slots
 - Check for orga/orga conflict and orga/participant conflicts
+- Add connection with compte clipper for ENS Ulm
 
 ## Version 3.0.0 - 2022-12-02 - Interludes 2023 Saclay